更新clangd和clang-tidy配置，移除不必要的检查项，添加新的检查项；修复llm_demo.cpp中的数据大小计算，确保类型安全；在rkllm.h中添加必要的头文件；更新online-recognizer.h以使用std::array替代原始数组。

+++ LLM_Voice_Flow/llm/rknn-llm/examples/DeepSeek-R1-Distill-Qwen-1.5B_Demo/deploy/src/llm_demo.cpp
@@ -221,8 +221,8 @@ void callback(RKLLMResult *result, void *userdata, LLMCallState state)
         ===============================================================================================================*/
         if (result->last_hidden_layer.embd_size != 0 && result->last_hidden_layer.num_tokens != 0)
         {
-            int data_size = result->last_hidden_layer.embd_size * result->last_hidden_layer.num_tokens * sizeof(float);
-            printf("\ndata_size:%d", data_size);
+            size_t data_size = static_cast<size_t>(result->last_hidden_layer.embd_size) * static_cast<size_t>(result->last_hidden_layer.num_tokens) * sizeof(float);
+            printf("\ndata_size:%zu", data_size);
             std::ofstream outFile("last_hidden_layer.bin", std::ios::binary);
             if (outFile.is_open())
             {
